Treat context like a global
There are a number of use cases - such as getting the current user -
where Resources and other objects need access to the context (in Rails,
the context is the controller). Resource#with_context made sure the
context was available for filter guards, for example.

This is now the same dev-facing API, but the context is set on
Thread.current rather than the Resource instance. This is because the
same context needs to be available to all nested resources, serializers,
etc. 'context' is more of a global concept, and this code treats it as
such.

The final product is you can do something like this in a Resource:

```ruby
def create(attributes)
  raise 'not authorized!' unless context.current_user.role == 'admin'
  super
end
```

...regardless of whether this resource is being created from its
corresponding endpoint, or side-posted 3 levels deep.