Better separate Rails-specific tests
There are now two sets of tests to run: library tests, and
rails-specific integration tests:

* `bin/rspec`
* `bin/appraisal bin/rspec spec/integration/rails`

This has always been roughly the case, but some things would be required
when they shouldn't be.

This *does* formally choose to depend on `active_support`. The sugar is
too nice to give up, and we'll eventually make heavy use of
`class_attribute`. Nothing else in Rails is required.

Also: we now always require the Null adapter, and the ActiveRecord
adapter is automatically required if ActiveRecord is defined (this
probably ends up in a Railtie).

Write-specific Rails tests will still fail for now.