Rename to Graphiti
Originally, this was a collection of gems (hence "Suite") specific to
the jsonapi.org specification (hence "JSONAPI"). We now support
rendering simple JSON and XML, and introduce first-class Resource
constructs instead of wiring-together various gems. We've outgrown the
name "JSONAPI Suite". Plus, JsonapiCompliable is a pain to type ;)

Instead, Graphiti: a stylish alternative to GraphQL.