Address real-world feedback
* Adapters now accept a resource instance in their constructor. This
allows greater introspection/flexibility, and access to runtime context.
* Improved error messages
* Ability to pass `single: true` to `filter` for filters that can only
accept a single value. The yielded value will not be an array in this
case. Present in schema + backwards-compatibility checked. Booleans
default this to `true`.
* Ability to pass `allow` and `reject` to `filter`. This way we can
disallow filtering on `nil` (which can impact performance), accomodate
enums, and boolean filters that can only ever be `true`. Present in
schema + backwards-compatibility checked.
* If using on-demand links, ensure `links=true` is appended to all urls.
* Avoid typecasting `nil` when appropriate.