Improve serialization typechecking performance
This was causing a hit to look up the attribute/type and perform the
logic. But we actually don't need that heavier-weight method for
serialization. The guard has already been run, and we already have the
type at definition-time. So, instead, register the serialization proc
with the type itself, rather than performing a runtime
lookup/validation.