Ensure attributes don't get excessively wrapped
There's a bunch of logic around serializer attributes: subclassing
resources, explicit serializers, defaults, overrides, etc. So we're
pretty dumb about saying "any time we're doing something that can affect
serialization, re-configure everything".

However, this caused the same attribute to be "wrapped" many times.
"Wrapping" is the process of taking an existing proc on the serializer
and applying type checking/coercion to it. It's needed when there is an
explicit serializer file/class.

Because we would re-fire the logic, we ended up wrapping and rewrapping
procs for no reason. This causes a performance penalty, particularly
with type-coercion firing over and over.

Instead, this commit keeps track of all attributes applied to the
serializer *from the Resource*. This way all the "re-configure
everything" logic can stay the same, but we'll only wrap procs when they
weren't applied by a resource (and thus need to be wrapped with
type-checking).