Accomodate PORO as_json
This is an extreme edge case, where we are processing a graphiti
request, the model is a PORO, and the developer manually does `model.as_json`
down the line. Because we add magic variables to the model instance,
they are processed and a StackLevelTooDeep error is thrown.

Instead, override the ActiveSupport method causing this behavior to
ignore the magic variables.