Lazily validate links when Rails autoloading
If we eagerly validate, and there is an error, the error is swallowed in
test mode - this is just the default autoloading behavior. The problem
is it then becomes very difficult to debug.

So instead: eagerly validate in all other scenarios, but lazily validate
if rails autoloading without eager class loading.