Omit placeholder relationships
The links_on_demand option is for those who care about payload size -
only render the links when specified in the request. This takes it to
the next level - avoid rendering the relationship unless sideloaded or
links are requested. If you are not using links_on_demand, this has no
effect.