Add persistence callbacks
These allow us to avoid overriding create/update directly and instead
rely on hooks:

* `before_attributes`: before the request attributes are assigned to the
model. Accepts attributes as argument.
* `after_attributes`: after the attributes have been assigned to the
but before save. Accepts the model as argument.
* `around_attributes`: Same as the above, but you must register with a
method name and `yield attributes` within it. Accepts attributes as
argument and the model instance is result of yield.
* `before_save`: after attributes have been assigned but before save.
Accepts model as argument.
* `after_save`: after the object has been persisted, but before that
transaction has been committed.
* `around_save`: Same as above, accepts the unpersisted model instance
and the persisted model instance is the result of yield.
* `before/after/around_destroy`: Same as all the above, accepts the
model instance.

See specs for details.