Add basic Resource & Attribute descriptions
Adding a `description` property to an attribute/extra_attribute
definition or setting `self.description` on any resource will cause the
that value to be output as part of the generated resource schema.
Additionally, if the `I18n` gem from Rails is loaded, descriptions can
be set via locale files, in the same way one might define error messages
or field labels.