Redefine total count when switching adapter
Total count is defined for all Resources by default. But if a resource
subclasses another and changes adapter, the total count will not use the
correct adapter because it was set before the adapter.

We should probably lazily grab the stat at runtime for this reason, but
this is the simple bugfix.