Set Query context to "show" during persistence actions
Due to some internal implementation needs (whch we want to clean up at
some point), the active record adapter has a bit of logic that is
dependent on whether the current request action is `create` or `update`.
This causes some slowness when dealing with sideload queries while
persisting a `many_to_many` relationship.  After we swtich over to the
sideload resolving step, we switch to a "reader" context to prevent this
from happening.