Only parse rails host/port in Server contexts
Rails::Server is only defined when running `bin/rails server` or
equivalents.  This prevents things from blowing up when in console or
rake task contexts while still allowing `ENV['HOST']` to be used.

Also adds a comment to explain the purpose of the code for people who
aren't sure why this is in their application.rb.