Fix sideposting of polymorphic resources
This is when the relationship is a normal has_many/belongs_to, but the
Resource is polymorphic and a specific jsonapi type is used. In this
case, we should be going through the correct child resource that matches
the jsonapi type for the logic.

Fixes https://github.com/graphiti-api/graphiti/issues/56