Optionally yield metadata on persistence hooks
Now, all persistence hooks will optionally yield metadata - a hash
containing things like the method, temp id, further relationships,
the calling model in the graph. We can continue hanging things on
metadata as we get use cases.

The main use case is referencing the "caller model" - the parent in the
graph. If we're saving an Employee and her Positions in a single
request:

```ruby
class PositionResource < ApplicationResource
  before_commit do |position, meta|
    meta[:caller_model] # => #<Employee>
  end
end
```

This is also true for override methods `#build`, `#assign_attributes`,
`#save`, and `#destroy` - all now accept `meta` as an optional argument,
so it is accessible when overriding.

A notable case is `destroy`. We do not currently handle nested destroys,
and the code path for destroying a top-level entity is different than
when sideposting a deletion. We accomodate both use cases here, but
notable.