Allow sideposting unwritable foreign keys
We want to allow sideposting/associating entities, but this should not
require exposing the foreign key as a writable attribute. Instead, we
infer the foreign key from configuration instead of the request, and
make sure this doesn't go through the normal writability check. If the
developer wants to prohibit writing the association, they can do
`has_many :positions, writable: false`.