Prevent query when empty has_many
We were already preventing a sideload query for a belongs_to when the
foreign key was nil. This wasn't done for has_many because we expect you
to always have a *primary* key.

This isn't the case for nonstandard primary keys explicitly specified.
So, let's make sure we prevent a query in this case.

Similarly, both have issues where the pk/fk is nil or unused, but the
parameters are being customized manually. In this case, we want to make
sure we don't accidentally *prevent* the query.