Make better use of build matrix in travis
Previously we were running a number of test scripts in a given build
run. We used the travis `rvm` setting to at least run a separate build
for each ruby version, but didn't do any more than that.  This made it
harder for contributors not as familiar with our testing setup
to identify the specific errors that happened.  This was particularly
problematic with things like linting, where failures could easily get
buried at the top of the file.

This change will also make it easier to skip certain tests for certain
combinations.  For example, with rails 6, ruby 2.4 is no longer
supported, so this will allow us to add rails 6 to our appraisal suite
while easily skipping test runs for it in ruby 2.4.