Wrap common sideload errors
If you already understand how Graphiti works, it's pretty easy to track
down why a sideload error is happening. But if that's not the case, you feel
kinda stuck with a random error. A good example is a nonstandard foreign key,
or forgetting to add an `attr_accessor` to your model.

This commit now wraps sideload errors, so we can provide context around
the place in the lifecycle the error is occurring:

* Parameter building
* Query Building
* Assignment

We'll still contain the original error message and backtrace in the
wrapped error message. This is also compatibile with Error#cause https://www.honeybadger.io/blog/nested-errors-in-ruby-with-exception-cause/

These errors usually only occur during development. Let's give
developers better information.