Don't link unexpected polymorphic type
If you try to sideload a `polymorphic_belongs_to`, and we encounter a
type not registered (and not in only/except), we'll raise a helpful
error. But if you aren't sideloading, and just generating a Link, and we
encounter an unknown type, you get a really confusing error.

This instead just omits the Link. I don't think we should blow up in
this case because the client isn't actually requesting anything here -
maybe the client only sideloads on data where the types are known.