Add "string_enum" type
When are dealing with an enum type, we don't want to accept anything not
in the list of values, nor do we want to niavely run them through the
string casting logic that exists in things like the activerecord
case-insensitive string filters.  When using the new `string_enum` type
in either your filters or your attributes, you will get an error if you
don't provide an `allow` option:

```ruby
\# raises Graphiti::Errors::MissingEnumAllowList
attribute :status, :string_enum

\# OK
attribute :status, :string_enum, allow: ['published', 'draft']
```

Note that this currently only works for filtering.  Eventually, I would
like to add the ability to prevent anything except for the allowlist
from being written automatically as well, but since users should be
handling this at their model or database layer anyway, I'm not worried
about it for this first version.