Add RequestValidator class
This is the first phase of work to allow us to provide better & more
actionable errors to API users who might not also be developers of the
API itself.  Previously when trying to write unwritable attributes &
relationships, on typecast problems, or trying to write an attribute
that doesn't exist, the request would blow up with a 500 and an
unhelpful error message.  The request validator takes on most of the
work of transforming and validating an incoming request for processing
all while rescuing any of the above errors and turning them into helpful
error messages.  Note that this is only currently intended to process
write payloads, but will be fully extendable to validate query
parameters for read requests too.

This doesn't actually integrate the validator into the code path, as I
want to get feedback on the design before doing that work in case
anything needs to change.  This will take over some of the
normalization responsibilities which currently live in
`Graphiti::Util::Persistence`.

This also adds a new "payload_path" key to the meta payload for the
deserialized/normalized request payload, which allows consumers to know
which path within the request json structure they are dealing with at
any given time. Here it is used for more useful error messages.