Defer serializer relationships when eager loading
To do endpoint validation, we register a proc in railtie.rb that checks
routes. For this to work, all resources and routes must be loaded. This
is a problem because the resource classes load before routes are
finished, and we want to do this at boot instead of runtime for
performance.

The solution is to defer the link validation until the app has loaded
*and* we've reloaded routes. This happens in railtie.rb, but it looks
like a straggler line would not defer, only under very specific
scenarios (when eager loading in production with polymorphic resources,
*and* a particular load order).

I've removed the straggler line to ensure this works as intended.