Fix remote sideload belongs_to links
These would point to the `show` action, as they do when linking locally.
However, remote belongs_to queries actually hit `index`.

You could make a case that we should fix this inconsistency, and I would
support a configuration option to do so. As it stands, that would A)
break backwards-compatibility B) possibly have the downside of alternate
remote logic based on relationship type C) hit caching and other
customizations related to `show` that might not be desired.

This PR just fixes the issue - correct link generation, and avoiding
errors when the `filter` param is customized.