Excape `match` operations to prevent DoS queries
Previously when doing a `match` operation against an activerecord string
field, we passed down the string directly and surrounded it with `%` on
either side, allowing a substring match of the column.  There was
nothing preventing a client from including additional `%` wildcard
characters in their query string to add more advanced substring
matching. The problem with this is that if an attacker provides a few
dozen wildcard characters, the sql engine can very quickly run into
processing problems and cause a Denial of Service against the
database/application.

This commit takes advantage of activerecord and arel's sql escaping
behavior to make this safe, at the tradeoff of not being able to pass
additional wildcard characters as part of queries.  Note that rails 4.x
has fewer APIs for this behavior and they aren't all public. Because
that is now an unsupported version of rails, this PR simply keeps the
less safe older behavior when that version of activerecord is detected.