Properly excape prefix and suffix queries
In #196 I added sanitization for `match` filters but I negelected
to handle `prefix` and `suffix` filters, which share similar logic.
This handles all three.  Additionally, issue #207 indicates an issue
with the `sanitize_sql_like` method not always being available on the
scope object. Wasn't able to reproduce myself, but this switches to
avoid using the method from scope directly and instead creates a utility
class which is extended with the appropriate sanitization methods.