Make context accessible when customizing relationship params
`context` can be useful when customizing params in a relationship. This
makes it accessible in the `params` block by adding it as the third
block variable:

```
class PostResource < ApplicationResource
  has_many :comments do
    params do |hash, _posts, context|
      hash[:filter] = { locale: context.locale }
    end
  end
end
```

The current workaround is to directly reference
`Graphiti.context[:object]` within the block, which is less desirable
because that private API could change.