Add option to deny empty filter values.
`deny_empty` is an option that can be set on a filter to raise an error
(a Graphiti::Errors::InvalidFilterValue error) if an empty value is
passed as the filter value.

It can be set at the resource level as a default and/or on individual
filters.

Default behaviour is setting this to false, which will not do anything
special with 'empty' filter values.

Setting on a Resource:
```
class BaseResource < Graphiti::Resource
  self.filters_deny_empty_by_default = true
end
```

Setting on an individual filter:
```
class UserResource < BaseResource
  filter :name, deny_empty: true
end
```