Add action to query
When performing a write activity on a resource and sideloading a
secondary, related resource, e.g. by `POST /resource/:id?include=related`
the namespace points to the original action for both the primary and
included resources.

This is useful because it provides context for the overall request, but
is incomplete and misleading for the action required on the included
resource and becomes problematic with authorization where the caller may
have write permission on the primary but only read access to the
sideload, meaning that the `Graphiti.context[:namespace]` cannot be
used in all cases.

Beacause a query is created for each resource, including those that are
sideloaded or sideposted, this change adds the action to the query
and sets it appropriately to `:read` when the action for that resource
is a read-type and prevents the action being incorrectly cascaded to
included sideloads.