Use joins/prelaod over includes for many_to_many
Ran into a strange edge case with activerecord's include behavior when
doing many_to_many relationship on a self-referential model. Still can't
quite figure out what the includes issue is, but I think it is actually
a bug in activerecord.  This commit switches to use a combination of
joins/preload when loading the assocation, which seems to do the trick.