Allow custom names for many_to_many filters
When creating a many_to_many relationship, graphiti adds a filter to the
foreign resource which is used when sideloading or provided as a link to
the relationship scope if link generation is enabled.  Sometimes,
however, the id of the field being looked up is an internal value and
might not be something you'd want to expose to the API itself.  This
allows us to pass a `filter_name` option when setting up many_to_many,
which it will use for the lookups.