Change filter_name => inverse_filter
Recent feedback from a user trying to implement the new filter_name
functionality was that they were putting the filter_name override on the
resource at the wrong side of the relationship. Adding inverse to the
argument name matches ActiveRecord's `inverse` argument to a
relationship definition and makes things more clear.

Additionally, I realized that overriding the filter from a link
generation perspective is desireable on has_many sideloads as well, so I
pulled this logic up a level into that sideload class.