Patch memory leak in Graphiti::Debugger.
The Debugger was adding data to the chunks Array via the
ActiveSupport::Notification handlers `on_data`, and `on_render`. When
the debugger is disabled, this Array is never flushed, and it is not
GC'd since it's referenced in a class instance variable.

This change returns early in both event handlers if the debugger
is not enabled. We also mark the utility methods in the debugger
as private so they cannot be accessed outside of the class (and add
data to chunks).