Avoid need for writable PK on update
There's some larger refactoring needed but this is the simplest possible
fix for the issue. There are three overall problems with the validator:

* Called for read operations, but meant to validate writes. I'm not sure
why this was ever the case, perhaps to eventually validate reads as
well.
* Called in multiple places (proxy and runner)
* References `params[:action]`, which couples us to Rails

So the ideal refactoring would be to only call the validator within
`ResourceProxy#save`, which addresses all the above.

The problem is one specific test looks at the metadata of a sideposted
resource (occasionally it's helpful to see things like "what parent
object is sideposting me"). Because read operations would go through the
validator and `id` would be typecast, this test now fails because it
sees `id` as a string instead of an integer, because it is no longer
being typecast.

I think I'm fine just updating the spec, but again this PR is to do the
simplest thing possible. So I instead added the very specific logic "if
validating id, and not a create operation, AND id is not writable, don't
attempt to typecast". The larger refactoring can come later.

Finally, the other future refactoring would be to avoid merging `id`
into `attributes` when not a create operation. Part of the issue here is
the deserializer is called even on reads (per above), which complicates
things further. So a good future approach would be to only call the
validator within `ResourceProxy#save`, then address deserialization
separately afterwards.