Ensure scope override is honored on update/destroy
If you were passing a scope override for an update or destroy operation,
that scope wouldn't be properly honored. This PR ensures an error will
be correctly thrown if no record is found.

The downside is we look up the record twice, now, but I think we should
get a fix in ASAP.