Bail out if trying to filter by a nil id
If we are trying to filter a resource by an id attribute but that
attribute is nil, we should raise an error. If we allow the resource
resolution to continue, the result will be an empty "where" clause and
taking the first result. This is a problem when sending `null` as the
relationship ID on a CREATE or UPDATE, as it would just associate
(and update!) the first record returned by the unscoped where clause.

We should probably have a check further up the stack as well in the
request validator or similar, but this at least puts a quick stop to the
issue and blows up these invalid requests while we take time to make a
more pleasant experience for API consumers.