Support polymorphic includes
Currently for polymorphic resources we just check if the include is
supported - if it is, render it. For instance in the sample app
`Epic` is a type of `Task` and it's the only one with the
`milestones` relationship. So `/tasks?include=milestones` works like
you'd think.

The issue is if both `Feature` and `Epic` support the `milestone`
relationship (or if all types do via the parent resource), you can't
conditionally request milestones just for epics. Enter this PR.

We introduce a special `on__<type>--<name>` syntax to conditionally
include based on type, e.g. `?include=on__epics--milestones`. We can
now support the use case. Note the syntax here is ugly, but we're tied
to a finite set of supported characters according to the JSON:API spec.

In practice I view this as a kind of internal API for supporting GQL
requests, so I'm only testing the hash renderer. I don't really want to
advertise it unless we get requests to support it for jsonapi.

Finally, note we're only doing this for includes, not fields. For fields
we can reference the jsonapi type as normal and this would only be an
issue if double-listing the same entities as multiple levels of the
graph. Maybe we support this one day, but for now I'd like to limit the
complexity/feature to just this case.