Remove previous guard when attribute overridden
Our actual use case is

```ruby
 # PositionResource
federated_type("Employee").has_many :positions
attribute :employee_id, :integer
```

Normally the first line with add the `employee_id` attribute with a
`readable: [:gql?]` guard, but the second line should override it.

This is also an issue in vanilla Graphiti, though very much an edge
case - if you are subclassing a resource and overriding an attribute
that was previously guarded but now is not, the guard should be removed.